login_and_signup_API_using_DRF/requirements.txt
Django==3.2.13
djangorestframework-api-key==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow

